Adds a common file
