+++ ce/instances.go
@@ -1,7 +1,6 @@
 package ce
 
 import (
-	"bytes"
 	"encoding/json"
 	"fmt"
 	"io/ioutil"
@@ -61,58 +60,6 @@ type Instance struct {
 		ObjectName string `json:"objectName"`
 		VendorName string `json:"vendorName"`
 	} `json:"transformationData"`
-}
-
-// Execute is a HTTP command that returns bytes, HTTP status, and a curl command
-func Execute(method, url, auth string) ([]byte, int, string, error) {
-	var bodybytes []byte
-	client := &http.Client{}
-	req, err := http.NewRequest(method, url, nil)
-	if err != nil {
-		fmt.Println("Can't construct request", err.Error())
-		os.Exit(1)
-	}
-	req.Header.Add("Authorization", auth)
-	req.Header.Add("Accept", "application/json")
-	req.Header.Add("Content-Type", "application/json")
-	curlCmd, _ := http2curl.GetCurlCommand(req)
-	curl := fmt.Sprintf("%s", curlCmd)
-	//log.Println(curl)
-	resp, err := client.Do(req)
-	if err != nil {
-		// unable to reach CE API
-		return bodybytes, -1, curl, err
-	}
-	bodybytes, err = ioutil.ReadAll(resp.Body)
-	defer resp.Body.Close()
-
-	return bodybytes, resp.StatusCode, curl, nil
-}
-
-// ExecuteWithBody is a HTTP command that returns bytes, HTTP status, and a curl command
-func ExecuteWithBody(method, url, auth string, requestbytes []byte) ([]byte, int, string, error) {
-	var bodybytes []byte
-	client := &http.Client{}
-	req, err := http.NewRequest(method, url, bytes.NewReader(requestbytes))
-	if err != nil {
-		fmt.Println("Can't construct request", err.Error())
-		os.Exit(1)
-	}
-	req.Header.Add("Authorization", auth)
-	req.Header.Add("Accept", "application/json")
-	req.Header.Add("Content-Type", "application/json")
-	curlCmd, _ := http2curl.GetCurlCommand(req)
-	curl := fmt.Sprintf("%s", curlCmd)
-	//log.Println(curl)
-	resp, err := client.Do(req)
-	if err != nil {
-		// unable to reach CE API
-		return bodybytes, -1, curl, err
-	}
-	bodybytes, err = ioutil.ReadAll(resp.Body)
-	defer resp.Body.Close()
-
-	return bodybytes, resp.StatusCode, curl, nil
 }
 
 // EnableElementInstanceEvents will enable or disable events on an Element Instance without requiring reauthentication
